Fix unbound variable in internal exit.
This would happen when exit was called with no argument.  Our internal
exit wouldn't pass along the exit code from the previous command and
instead would blow up because it didn't receive a $1.  Luckily almost
all of our code is checked carefully to not call exit without a
parameter so this bug might not actually affect anyone.