Enhance archive_create to take multiple sources.
This commit provides common abstraction in archive_create so that you
can seamlessly provide multiple source paths to be archived. The source
paths can also now be files or directories. This is implemented
consistently across all archive formats in the manner implemented by
mksquashfs. Namely, a single directory argument will add its contents
recursively. If multiple arguments are given each is the name of a the
top-level entry in the archive.