Updated to allow color specifications
Moved things over so the commonly used colors are specified by a variable
instead of hard-coded, this allows you to modify the color schemes via using
either /etc/bashutilsrc or ~/.bashutilsrc (or some combination), which are
automatically sourced if they exist. If neither file exists, the default
(current) color scheme is used.