Make netselect faster against a single host.
This gives netselect an option(--count) that can specify how many times
to ping each host during the test.  Historically this was 10 times, and
that is retained as the default _except_ when there is only one host to
test.  In that case, the default is to perform a single ping.

An argument could be made for simply returning the only host and not
pinging it at all, but it seems better to verify that we have at least
_some_ connectivity to the single host before doing so.