Shadow bash's trap builtin with a function
Bash's trap builtin has the terribly annoying habit of resetting the
current shell's DEBUG and ERR traps when it creates a new trap for the
first time in a shell.  Even when the proper options are set to cause
those DEBUG and ERR traps to be inherited.

This creates a trap function that corrects for this issue.

Also creates a new "truncate" function to ease truncation of long
strings that end up in user output.

Adds several tests, particularly around:
- Messaging produced from error traps.
- Traps themselves.
- The new truncate function.