Add new overlayfs_commit operation.
This function will commit all pending changes in the overlayfs write layer
back down into the lowest read-only layer and then unmount the overlayfs
mount. The intention of this function is that you should call it when
you're completely done with an overlayfs and you want its changes to
persist back to the original archive.