Make opt_parse stop tickling a bash bug
This bash bug affects at least 4.3.11 (which is in ubuntu 14.04) and
4.3.30 (which is in ubuntu 15.04).  Basically, the trouble is that foo
behaves differently than bar.

    foo() { declare -a "ARR=(a)" ; declare -p ARR ; }
    bar() { declare -a ARR=(a) ; declare -p ARR ; }

Whether foo should work in general may be questionable, but it does work
in bash 4.2 and also in newer 4.3s like 4.3.42 from ubuntu 15.10.  In
any case, both cases work outside of a function even in the broken
versions of bash.

This thread on the bash developer mailing list seems to describe the
issue best, and in this particular message Chet Ramey (bash author)
attached a bash patch that could fix it.

Luckily, the bashutils changes in this commit change to avoid tickling
the bug and I think it simplifies things slightly in the process.