overlayfs: detect bootable ISOs better
Some versions of the 'file' program print a string that includes
"(bootable)" when presented with a bootable ("hybrid") ISO image, and
some do not.

This commit makes the overlayfs_commit function be more accepting of
different ways of expressing that an ISO is bootable.  Fixes EMBER-866:
fix neon-ember iso boot.